Support and use rounded corners for widgets
Follows Blender's code pretty closely. Did some additional changes that
were either needed or simply made things a bit nicer.

+++ bwidgets/util/Point.cc
@@ -2,7 +2,7 @@
 
 using namespace bWidgets;
 
-Point::Point(int x, int y) :
+Point::Point(float x, float y) :
     x(x), y(y)
 {
